🧪 Consolidate and improve Proxy Headers Middleware tests
- Consolidated fragmented proxy test files into a single, comprehensive suite in `tests/test_proxy_headers.py`.
- Updated `proxy_headers` factory in `xyra/middleware/proxy_headers.py` to support string input and default to '127.0.0.1'.
- Added robust tests for scheme, host, and port forwarding.
- Included security tests for IP spoofing and DoS mitigation.
- Removed redundant test files: `tests/test_middleware_proxy_headers.py`, `tests/test_proxy_headers_extensions.py`, `tests/test_proxy_dos_fix.py`, and `tests/test_security_proxy_dos.py`.
